backend/requirements.txt
fastapi
redis
types-redis
uvicorn
numpy
scipy
opencv-python
pillow